requirements.txt
flask
biopython
scikit-learn
numpy
pandas
matplotlib
seaborn
joblib
xgboost
fpdf2
reportlab
pyttsx3
plotly
gtts
opencv-python
scipy
face-recognition
Pillow
